Introduce test fixture
