+++ discovery-service/src/test/java/org/zowe/apiml/discovery/functional/HttpsSecuredEndpointsTest.java
@@ -11,8 +11,12 @@
 package org.zowe.apiml.discovery.functional;
 
 import org.apache.http.HttpStatus;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.test.context.ActiveProfiles;
+import org.zowe.apiml.util.config.SslContext;
+import org.zowe.apiml.util.config.SslContextConfigurer;
 
 import static io.restassured.RestAssured.given;
 import static org.hamcrest.core.Is.is;
@@ -25,10 +29,27 @@ public class HttpsSecuredEndpointsTest extends DiscoveryFunctionalTest {
         return "https";
     }
 
+    @Value("${server.ssl.keyStorePassword:password}")
+    private char[] keystorePassword;
+
+    @Value("${server.ssl.keyStore}")
+    private String keystore;
+
+    private String clientKeystore = "../keystore/client_cert/client-certs.p12";
+
+    @Override
+    @BeforeEach
+    void setUp() throws Exception {
+        SslContextConfigurer configurer = new SslContextConfigurer(
+            keystorePassword,
+            clientKeystore,
+            keystore);
+        SslContext.prepareSslAuthentication(configurer);
+    }
+
     @Test
     void testEurekaEndpoints_whenProvidedCertificate() throws Exception {
-        //RestAssured.config = RestAssured.config().sslConfig(getConfiguredSslConfig());
-        given()
+        given().config(SslContext.clientCertApiml)
             .when()
             .get(getDiscoveryUriWithPath("/eureka/apps"))
             .then()
