+++ discovery-service/src/test/resources/application.yml
@@ -25,6 +25,7 @@ server:
     address: ${apiml.service.ipAddress}
     port: ${apiml.service.port}
 
+
 eureka:
     instance:
         hostname: ${apiml.service.hostname}
@@ -46,3 +47,23 @@ management:
     health:
         defaults:
             enabled: false
+---
+spring:
+    profiles: https
+
+server:
+    ssl:
+        enabled: true
+        clientAuth: want
+        protocol: TLSv1.2
+        enabled-protocols: TLSv1.2
+        ciphers: TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256,TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256,TLS_ECDHE_RSA_WITH_AES_256_GCM_SHA384,TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384,TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA256,TLS_ECDHE_ECDSA_WITH_AES_128_CBC_SHA256,TLS_ECDHE_ECDSA_WITH_AES_256_CBC_SHA384
+        keyStoreType: PKCS12
+        trustStoreType: PKCS12
+        trustStoreRequired: true
+        keyAlias: localhost
+        keyPassword: password
+        keyStore: ../keystore/localhost/localhost.keystore.p12
+        keyStorePassword: password
+        trustStore: ../keystore/localhost/localhost.truststore.p12
+        trustStorePassword: password
