+++ integration-tests/src/test/java/org/zowe/apiml/integration/authentication/schemes/X509SchemeTest.java
@@ -14,8 +14,8 @@ import org.junit.jupiter.api.*;
 import org.zowe.apiml.util.TestWithStartedInstances;
 import org.zowe.apiml.util.categories.DiscoverableClientDependentTest;
 import org.zowe.apiml.util.categories.X509Test;
+import org.zowe.apiml.util.config.ItSslConfigFactory;
 import org.zowe.apiml.util.config.SslContext;
-import org.zowe.apiml.util.config.SslContextConfigurer;
 import org.zowe.apiml.util.http.HttpRequestUtils;
 
 import java.net.URI;
@@ -36,7 +36,7 @@ class X509SchemeTest implements TestWithStartedInstances {
 
     @BeforeAll
     static void init() throws Exception {
-        SslContext.prepareSslAuthentication(SslContextConfigurer.integrationTests());
+        SslContext.prepareSslAuthentication(ItSslConfigFactory.integrationTests());
         URL = HttpRequestUtils.getUriFromGateway(X509_ENDPOINT);
     }
