+++ integration-tests/src/test/java/org/zowe/apiml/integration/zos/ServicesInfoTest.java
@@ -59,7 +59,7 @@ class ServicesInfoTest implements TestWithStartedInstances {
 
     @BeforeAll
     public static void init() throws Exception {
-        SslContext.prepareSslAuthentication(SslContextConfigurer.integrationTests());
+        SslContext.prepareSslAuthentication(ItSslConfigFactory.integrationTests());
     }
